Use map-filter in strip_punc
This takes the loops into C-land (in CPython, at least), which is
generally a good idea, although I haven't profiled anything so probably
I'm the root of all evil. But then again I also do quite like the
functional style of using map and filter over comprehensions, so it's
not just an optimisation.